# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Memory Threads API - AI-powered memory engine for files
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List all memories
      parameters:
        - in: query
          name: instanceId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, ready, error]
        - in: query
          name: fileType
          schema:
            type: string
        - in: query
          name: people
          schema:
            type: string
        - in: query
          name: topics
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryListResponse"
    delete:
      operationId: deleteAllMemories
      tags: [memories]
      summary: Delete all memories (optionally scoped to an instance)
      parameters:
        - in: query
          name: instanceId
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAllResponse"
  /memories/upload:
    post:
      operationId: uploadMemory
      tags: [memories]
      summary: Upload one or more files (including .zip/.7z archives) to create memories
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MemoryFileUpload"
      responses:
        "201":
          description: Memories created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryUploadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a single memory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a single memory by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /memories/{id}/related:
    get:
      operationId: getRelatedMemories
      tags: [memories]
      summary: Get memories related to a given memory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
...
components:
  schemas:
...
    MemoryFileUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          description: The file to upload (binary)
        instanceId:
          type: integer
          description: Instance to assign the memory to
...
    Memory:
      type: object
      required:
        - id
        - originalName
        - fileType
        - fileSize
        - status
        - uploadedAt
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        instanceId:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, ready, error]
        confidence:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            type: string
        organizations:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            type: string
        dates:
          type: array
          items:
            type: string
        tasks:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: ["string", "null"]
          format: date-time
        errorMessage:
          type: ["string", "null"]
    MemoryListResponse:
      type: object
      required:
        - memories
        - total
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
        total:
          type: integer
...
    MemoryUploadResponse:
      type: object
      required:
        - memories
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
...
    DeleteResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
    DeleteAllResponse:
      type: object
      required:
        - success
        - deletedCount
      properties:
        success:
          type: boolean
        deletedCount:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string